backend/pyproject.toml
[project]
name = "twin-analytics-backend"
version = "1.0.0"
description = "Backend API for AI Twin Analytics Dashboard"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "faker>=20.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
